Use size_t instead of a signed 32 bit integer to compute address in pcl::CropBox<pcl::PointCloud2> (fixes #3417)

+++ filters/src/crop_box.cpp
@@ -78,8 +78,8 @@ pcl::CropBox<pcl::PCLPointCloud2>::applyFilter (PCLPointCloud2 &output)
   for (size_t index = 0; index < indices_->size (); ++index)
   {
     // Get local point
-    int point_offset = ((*indices_)[index] * input_->point_step);
-    int offset = point_offset + input_->fields[x_idx_].offset;
+    size_t point_offset = static_cast<size_t>((*indices_)[index]) * input_->point_step;
+    size_t offset = point_offset + input_->fields[x_idx_].offset;
     memcpy (local_pt.data (), &input_->data[offset], sizeof (float)*3);
 
     // Check if the point is invalid
@@ -168,8 +168,8 @@ pcl::CropBox<pcl::PCLPointCloud2>::applyFilter (std::vector<int> &indices)
   for (size_t index = 0; index < indices_->size (); index++)
   {
     // Get local point
-    int point_offset = ((*indices_)[index] * input_->point_step);
-    int offset = point_offset + input_->fields[x_idx_].offset;
+    size_t point_offset = static_cast<size_t>((*indices_)[index]) * input_->point_step;
+    size_t offset = point_offset + input_->fields[x_idx_].offset;
     memcpy (local_pt.data (), &input_->data[offset], sizeof (float)*3);
 
     // Transform point to world space
